# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Trivial Scanner Dashboard API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/claim/{client_name}": {
      "post": {
        "tags": [
          "Client Server"
        ],
        "summary": "Claim Client",
        "description": "Generates an access token for provided *NEW* client name.\nClient names must be unique, if the corresponding registration token was lost a new client and token must be created.",
        "operationId": "claim_client_claim__client_name__post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Client Name",
              "type": "string"
            },
            "name": "client_name",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Trivialscan-Version",
              "type": "string"
            },
            "name": "x-trivialscan-version",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            },
            "name": "authorization",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Trivialscan-Account",
              "type": "string"
            },
            "name": "x-trivialscan-account",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientInfo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          },
          "400": {
            "description": "The client name must not be an email address"
          },
          "401": {
            "description": "Authorization Header was sent but something was not valid (check the logs), likely signed the wrong HTTP method or forgot to sign the base64 encoded POST data"
          },
          "403": {
            "description": "Authorization Header was not sent, or dropped at a proxy (requesters issue) or the CDN (that one is our server misconfiguration)"
          },
          "409": {
            "description": "This client name has already been registered"
          },
          "500": {
            "description": "An unhandled error occurred during an AWS request for data access"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/{client_name}": {
      "post": {
        "tags": [
          "Client Server"
        ],
        "summary": "Auth Client",
        "description": "Authenticates the generated access token and client",
        "operationId": "auth_client_auth__client_name__post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Client Name",
              "type": "string"
            },
            "name": "client_name",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Trivialscan-Version",
              "type": "string"
            },
            "name": "x-trivialscan-version",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            },
            "name": "authorization",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Trivialscan-Account",
              "type": "string"
            },
            "name": "x-trivialscan-account",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientInfo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckToken"
                }
              }
            }
          },
          "400": {
            "description": "The client name must not be an email address"
          },
          "401": {
            "description": "Authorization Header was sent but something was not valid (check the logs), likely signed the wrong HTTP method or forgot to sign the base64 encoded POST data"
          },
          "403": {
            "description": "Authorization Header was not sent, or dropped at a proxy (requesters issue) or the CDN (that one is our server misconfiguration)"
          },
          "500": {
            "description": "An unhandled error occurred during an AWS request for data access"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clients": {
      "get": {
        "tags": [
          "Client Server"
        ],
        "summary": "Retrieve Clients",
        "description": "Retrieves a collection of your clients",
        "operationId": "retrieve_clients_clients_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            },
            "name": "authorization",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Trivialscan-Account",
              "type": "string"
            },
            "name": "x-trivialscan-account",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Retrieve Clients Clients Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientRedacted"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No client name exists for this account"
          },
          "401": {
            "description": "Authorization Header was sent but something was not valid (check the logs), likely signed the wrong HTTP method or forgot to sign the base64 encoded POST data"
          },
          "403": {
            "description": "Authorization Header was not sent, or dropped at a proxy (requesters issue) or the CDN (that one is our server misconfiguration)"
          },
          "500": {
            "description": "An unhandled error occurred during an AWS request for data access"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/activate/{client_name}": {
      "get": {
        "tags": [
          "Client Server"
        ],
        "summary": "Activate Client",
        "description": "Activate a activated client",
        "operationId": "activate_client_activate__client_name__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Client Name",
              "type": "string"
            },
            "name": "client_name",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            },
            "name": "authorization",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Trivialscan-Account",
              "type": "string"
            },
            "name": "x-trivialscan-account",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientRedacted"
                }
              }
            }
          },
          "204": {
            "description": "Client name does not exists for this account"
          },
          "401": {
            "description": "Authorization Header was sent but something was not valid (check the logs), likely signed the wrong HTTP method or forgot to sign the base64 encoded POST data"
          },
          "403": {
            "description": "Authorization Header was not sent, or dropped at a proxy (requesters issue) or the CDN (that one is our server misconfiguration)"
          },
          "500": {
            "description": "An unhandled error occurred during an AWS request for data access"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/deactivated/{client_name}": {
      "get": {
        "tags": [
          "Client Server"
        ],
        "summary": "Deactivated Client",
        "description": "Deactivate a client",
        "operationId": "deactivated_client_deactivated__client_name__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Client Name",
              "type": "string"
            },
            "name": "client_name",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            },
            "name": "authorization",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Trivialscan-Account",
              "type": "string"
            },
            "name": "x-trivialscan-account",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientRedacted"
                }
              }
            }
          },
          "204": {
            "description": "Client name does not exists for this account"
          },
          "401": {
            "description": "Authorization Header was sent but something was not valid (check the logs), likely signed the wrong HTTP method or forgot to sign the base64 encoded POST data"
          },
          "403": {
            "description": "Authorization Header was not sent, or dropped at a proxy (requesters issue) or the CDN (that one is our server misconfiguration)"
          },
          "500": {
            "description": "An unhandled error occurred during an AWS request for data access"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/client/{client_name}": {
      "delete": {
        "tags": [
          "Client Server"
        ],
        "summary": "Delete Client",
        "description": "Deletes a specific MemberProfile within the same account as the authorized requester",
        "operationId": "delete_client_client__client_name__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Client Name",
              "type": "string"
            },
            "name": "client_name",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            },
            "name": "authorization",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Trivialscan-Account",
              "type": "string"
            },
            "name": "x-trivialscan-account",
            "in": "header"
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "The client name cannot be an email address"
          },
          "401": {
            "description": "Authorization Header was sent but something was not valid (check the logs), likely signed the wrong HTTP method or forgot to sign the base64 encoded POST data"
          },
          "403": {
            "description": "Authorization Header was not sent, or dropped at a proxy (requesters issue) or the CDN (that one is our server misconfiguration)"
          },
          "500": {
            "description": "An unhandled error occurred during an AWS request for data access"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CheckToken": {
        "title": "CheckToken",
        "type": "object",
        "properties": {
          "version": {
            "title": "Version",
            "type": "string"
          },
          "session": {
            "$ref": "#/components/schemas/MemberSessionRedacted"
          },
          "client": {
            "$ref": "#/components/schemas/ClientRedacted"
          },
          "account": {
            "$ref": "#/components/schemas/MemberAccountRedacted"
          },
          "member": {
            "$ref": "#/components/schemas/MemberProfileRedacted"
          },
          "authorisation_valid": {
            "title": "HMAC Signature validation",
            "type": "boolean",
            "description": "Provides verifiable proof the client has possession of the Registration Token (without exposing/transmitting the token), using SHA256 hashing of the pertinent request information",
            "default": false
          },
          "ip_addr": {
            "title": "Ip Addr",
            "type": "string",
            "description": "Source IP Address"
          },
          "user_agent": {
            "title": "User Agent",
            "type": "string",
            "description": "Source HTTP Client"
          }
        }
      },
      "Client": {
        "title": "Client",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/MemberAccount"
          },
          "client_info": {
            "$ref": "#/components/schemas/ClientInfo"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "cli_version": {
            "title": "Cli Version",
            "type": "string"
          },
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "ip_addr": {
            "title": "Ip Addr",
            "type": "string",
            "format": "ipvanyaddress"
          },
          "user_agent": {
            "title": "User Agent",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "integer"
          },
          "active": {
            "title": "Active",
            "type": "boolean",
            "default": false
          }
        }
      },
      "ClientInfo": {
        "title": "ClientInfo",
        "type": "object",
        "properties": {
          "operating_system": {
            "title": "Operating System",
            "type": "string"
          },
          "operating_system_release": {
            "title": "Operating System Release",
            "type": "string"
          },
          "operating_system_version": {
            "title": "Operating System Version",
            "type": "string"
          },
          "architecture": {
            "title": "Architecture",
            "type": "string"
          }
        }
      },
      "ClientRedacted": {
        "title": "ClientRedacted",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/MemberAccount"
          },
          "client_info": {
            "$ref": "#/components/schemas/ClientInfo"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "cli_version": {
            "title": "Cli Version",
            "type": "string"
          },
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "ip_addr": {
            "title": "Ip Addr",
            "type": "string",
            "format": "ipvanyaddress"
          },
          "user_agent": {
            "title": "User Agent",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "integer"
          },
          "active": {
            "title": "Active",
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "MemberAccount": {
        "title": "MemberAccount",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "display": {
            "title": "Display",
            "type": "string"
          },
          "primary_email": {
            "title": "Primary Email",
            "type": "string",
            "format": "email"
          },
          "billing_email": {
            "title": "Billing Email",
            "type": "string",
            "format": "email"
          },
          "api_key": {
            "title": "Api Key",
            "type": "string"
          },
          "ip_addr": {
            "title": "Ip Addr",
            "type": "string",
            "format": "ipvanyaddress"
          },
          "user_agent": {
            "title": "User Agent",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "integer"
          },
          "billing": {
            "$ref": "#/components/schemas/Billing"
          }
        }
      },
      "MemberAccountRedacted": {
        "title": "MemberAccountRedacted",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "display": {
            "title": "Display",
            "type": "string"
          },
          "primary_email": {
            "title": "Primary Email",
            "type": "string",
            "format": "email"
          },
          "billing_email": {
            "title": "Billing Email",
            "type": "string",
            "format": "email"
          },
          "api_key": {
            "title": "Api Key",
            "type": "string"
          },
          "ip_addr": {
            "title": "Ip Addr",
            "type": "string",
            "format": "ipvanyaddress"
          },
          "user_agent": {
            "title": "User Agent",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "integer"
          },
          "billing": {
            "$ref": "#/components/schemas/Billing"
          }
        }
      },
...
      "MemberSessionRedacted": {
        "title": "MemberSessionRedacted",
        "required": [
          "session_token"
        ],
        "type": "object",
        "properties": {
          "member": {
            "$ref": "#/components/schemas/MemberProfile"
          },
          "session_token": {
            "title": "Session Token",
            "type": "string"
          },
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "ip_addr": {
            "title": "Ip Addr",
            "type": "string",
            "format": "ipvanyaddress"
          },
          "user_agent": {
            "title": "User Agent",
            "type": "string"
          },
          "browser": {
            "title": "Browser",
            "type": "string"
          },
          "platform": {
            "title": "Platform",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "integer"
          },
          "map_svg": {
            "title": "Map Svg",
            "type": "string"
          },
          "current": {
            "title": "Current",
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      },
...
    }
  }
}